Refactor to use Settings for RenderAPI
It was a risk to have the RenderAPI setting directly within the RenderContext as it could lead to inconsistent states. Moved RenderAPI into core::Settings to ensure consistency across the engine. RenderAPI is now retrieved from Settings throughout the engine. Adjusted the create method in RenderContext to no longer require Settings as a parameter, since it can access these directly.