Renamed RenderContext::RenderAPI to core::Settings::RenderAPI in ClearCommand.cpp to ensure consistency across the codebase. Fixed parallel-hashmap path in the CMakeLists.txt file to resolve build errors.
Also, added necessary checks in RenderPipeline.hpp and refactored the code to make it clearer and type-safe. These changes provide better control and make the pipeline easier to use and understand.

The changes will hence improve code readability and ensure correct usage of RenderCommands. The addition of RenderPipeline to the GameLoop streamlines the render process and application for developers.