Refactor hexen::editor to enhance code readability and accessibility
Modified hexen::editor's EditorGameLoop and EditorGUI classes to enhance code readability and maintainability. Added RenderContext as a shared pointer parameter to both EditorGameLoop and EditorGUI constructors to setup render context for graphics. This makes it more explicit that they require the RenderContext and improve the cohesion of the classes. In EditorGameLoop, changed the code to use the new RenderPipeline methods instead of Window's clear method to improve consistency with the graphics updates.