Refactor graphics textures to use shared pointers
Changed graphics textures in the AssetsIcon from unique pointers to shared pointers. This modification enhances the flexibility and safety by allowing multiple parts of the program to access the same textures comfortably. It also enables more reliable control over the lifetime of the textures. Besides, the GLTexture2D class was swapped to the more general Texture class, making the code more adaptable and decreasing the dependency on OpenGL-specific code.