Refactor FramebufferWindow to use RenderPipeline
Significant changes regarding framebuffer handling in FramebufferWindow. Previously direct GL calls and method calls on GL objects were responsible for setting up and using framebuffer.

The codebase has been refactored to leverage the RenderPipeline, providing an abstraction layer and making the code cleaner and easier to maintain. Now, a FramebufferCommand is generated to replace explicit OpenGL framebuffer bindings, giving a simple interface to render to texture in FramebufferWindow.cpp.

The clear method has also been refactored to utilise ClearCommand instead of explicitly setting the clear color and clearing the buffer, which makes clearing operations consistent throughout the application.

These changes also aid in the potential for switching to another rendering API in the future with minimal adjustments in application code.