Adjust ShaderProgram create function to constant parameter
This diff modifies the ShaderProgram create function as well as its implementations and usages to receive a constant reference to the shader paths vector. Formerly, the function was receiving a non-constant reference. This change is to ensure the vector of paths that the caller passes in will not be accidentally altered in the process, thus enhancing the safety and reliability of the code. Also, the process of reading in shader files and compiling/linking shader programs has improved: instead of doing these operations for every single shader file, these operations are now done collectively after all shader files are read in.