Add setData function to GLVertexBuffer class
A setData function was added to the GLVertexBuffer class in engine/src/graphics/buffers/GL to allow setting of data to the GLVertexBuffer in the GL of the graphics engine. This function binds the buffer to the GL_ARRAY_BUFFER target and creates and initializes a buffer object's data store. This change provides the ability to directly manipulate data in the GLVertexBuffer for customizable graphic rendering.