Add triangles rendering function
Introduced a new function, drawTriangles, to facilitate the rendering of triangles, supporting different rendering APIs - No API, OpenGL, Vulkan, and DirectX12. Currently, only OpenGL rendering is implemented, others are placeholders. This function also adds to the profile for performance monitoring. Corresponding calls in GLDrawCalls are also added. This new functionality increases the flexibility and capacity of the graphics engine.