"Add support for texture arrays and improve texture filter conversion
Added TextureArray class to support 2D texture arrays for more efficient texture storage and sampling. Furthermore, the position of textureFilterToGLTextureFilter function is moved from the GLTexture2D.cpp to GLTexture2D.hpp for direct and easier accessibility. This enhancement ensures the improved handling of textures in the engine and promotes better rendering performance."