"Refactor DebugGridComponent for greater customization
Refactored DebugGridComponent to allow for customization of the grid's color and shader programs. DebugGridComponent now utilizes DrawGridCommand, to encapsulate the drawing behavior and make the application of color and shader programs more intuitive. This change enhances the flexibility and usability of the grid rendering process, simplifying customization and improving visual representation."