Change GridComponent member from unique to shared pointer
Switching 'grid' from a unique pointer to a shared pointer in gridComponent class to enable multiple ownership. This is in line with the refactor towards a dedicated EditorRenderSystem, where several components could potentially own and modify the same grid. This change enhances efficiency and allows more streamlined control of rendering components.