Optimize rendering process by centralizing grid control
Strip down EditorGameLoop.cpp by removing DebugGrid management logic from it and embedding it into SceneWindow constructor for better performance and efficiency. The change had to be made because we wanted to consolidate grid control into EditorRenderSystem for improved rendering control and component management. An added bonus is the system's multicomponent compatibility.