Add support for GridComponent in EditorRenderSystem
Added support for GridComponent in EditorRenderSystem. This was implemented for development purposes where having a visual grid in the editor helps in object placement and scene manipulation. The change was facilitated by adding methods to register, retrieve, and release GridComponent respectively. A static assertion has been included to ensure the type compatibility on the target system. This becomes imperative in case a component of type T does not exist in the system. Through this, an increase in debugging and design efficiency is expected.