Commit message:
Add AssetStorage for efficient asset management

Implemented an asset management system through the introduction of the AssetsStorage class. This class uses a Parallel Hash Map to allow for effective and fast storage and retrieval of various game assets. Moreover, AssetsStorage handles the root directory for all assets which will simplify asset access throughout the game engine. This update will pave the way for a more streamlined and organized approach to managing game assets.