Add createAsset method in AssetsStorage class
Added a generic function `createAsset' to create and manage assets in the `AssetsStorage` class. This function creates a new asset if the given asset does not exist, else it saves the existing asset. This allows robust and thread-safe handling of assets. It throws a static assertion error if the type 'T' is not inherited from `IAsset` interface, ensuring type validation at compile-time.