Add default asset storage functionality in AssetsStorage
Introduced a new function `addDefaultStorage` in AssetsStorage class to provide a simplified way to add a default asset storage. It uses `addAssetsStorage` function internally with "default-storage" as the storage name. This is useful when there is a need to set a common storage for assets which can be used as a fallback option. The function takes the root directory of the assets as the parameter.

Additional modifications include the update of comments for the `addAssetsStorage` function in the `AssetsStorage.hpp` file to reflect the changes and the addition of the `AssetsHelper` class providing utility methods for loading and creating assets.