Add IAsset interface for in-memory asset representation
Introduced an IAsset interface as a standard for creating Asset objects within a game engine. The interface provides a set of pure virtual methods to load, save and access general info about an asset. This interface should ensure a consistent API for all possible asset classes, making them easier to manage within containers, and at the same time, it provides flexibility for developers to create their propio asset subclasses. This change lays the foundation of the asset loading and management system in the engine.