Refactor ShaderFile class to ShaderAsset implementing IAsset
This commit refactors the existing 'ShaderFile' class and turns it into a 'ShaderAsset' class implementing the 'IAsset' interface. Now, ShaderAsset can be handled more consistently alongside other assets in the game engine system. 'ShaderFile' was essentially serving as a shader asset and renaming it to 'ShaderAsset' better represents its purpose. It introduces new methods to save, load, and provide essential information about the shader assets which are in compliance with the 'IAsset' interface. Additionally, shader loading from and saving to disk operations are also refactored for a more abstract and flexible approach. The change is part of the broader effort to improve asset management in the game engine.