Add AssetsStorage to Application and update paths
Added an AssetsStorage static method in the Application class. The change enables sharing of the assets storage across different classes without the need to instantiate an Application class. The paths in AssetsStorage's addAssetsStorage method have also been updated for EditorData and EngineData folders. This change provides more efficient access and organization of game assets, and ties the lifecycle of assets storage to the Application, and not an individual instance of the Application class.