Refactor ShaderAsset to support retrieval of shader text
In ShaderAsset classes, ShaderProgram import is moved and a new getShaderText method is added, returning const char pointer to the stored shader content. In the load function, the shader text is extracted and stored upon loading and returned later via getShaderText. This allows for a more flexible, efficient asset management allowing direct access to the shader text. Binary mode is now used when loading file to include all non-text characters.