Update shader file naming in CMakeLists
The Shader files have been renamed to ShaderAsset from ShaderFile in the 'engine/src/graphics/shaders/CMakeLists.txt' file. This was done to better reflect the functionalities of the file which include supporting the retrieval of shader text, as well as streamlining our asset management system. The change allows for a more flexible and efficient way of accessing the stored shader content, including all non-text characters.