Correct image orientation in GL texture array and 2D texture
This commit includes calling the 'flip' method on 'imageAsset' instances in the GLTextureArray and GLTexture2D classes. The change was necessary because the image texture assets were displayed upside down in the OpenGL context. By flipping the images during the texture generation phase, this commit ensures the correct orientation and presentation of graphics assets.