Add option to flip texture in Texture2D and GL classes
Added a 'flipImage' parameter to Texture2D, GLTexture2D, and GLTextureArray classes methods. This parameter is defaulted to false. If set to true, the image asset is flipped. This functionality was added to provide control over the orientation of the texture when using Texture2D and GLTexture2D classes, as certain rendering or modeling scenarios may require the image asset to be loaded 'flipped'. Also adopted changes in their respective implementation files. This is a non-breaking change as the parameter will default to false if not provided.