Enhancing asset handling and project structuring
This commit improves the asset handling and project structuring processes. In the case of AssetStorage, the setAssetsRootDirectory function is modified to include an optional parameter for moving all current assets to a new directory if the flag is set. The addition of a method for getting the default AssetStorage instance has also been made for streamlined asset management across the project.

The diff also involves changes to the way file paths are addressed and managed in the Project class. The use of std::filesystem is now more prevalent for safer path manipulations in contrast to concatenated strings.

Additionally, the HEXEN_ASSERT macro has seen an update to now print to std::cerr instead of fputs further improving the error reporting process.

Finally, an Application::getName function is declared and implemented to provide a standard way to retrieve the application name. This aids in maintaining uniformity when referring to the application in different parts of the code.