Add file extension retrieval in ShaderAsset class
Added a new member function in the ShaderAsset class to provide a way of retrieving the file extension of a ShaderAsset. This enhancement aids in identifying the type of ShaderAsset. New function named `getExtension` returns a string view representing the extension of the ShaderAsset file. This change increases code readability and enhances functionality.