Remove Window's clear function and SDL_image dependency
Removed the clear function in Window class and the SDL_image dependency in GLTexture2D. The clear functionality is no longer needed due to redesign in the rendering pipeline. The SDL_image library is replaced with STBImage for lightweight code footprint and as an update of our project's dependency policy which encourages reducing third-party dependencies. Some SDL_image-based icon loading code in Window class was also adapted to STBImage.