Add CheckerboardQuad component to manage checkerboard rendering
A new component named CheckerboardQuad has been added for managing checkerboard rendering in the game engine. This was included to support debugging in graphics applications. As part of this addition, 'EditorRenderSystem.cpp' and 'EditorRenderSystem.hpp' were updated to include this new component. In an effort to improve efficiency, the COMPONENTS_CONTAINER_SIZE has also been reduced from 100 to 10. The CheckerboardQuadComponent files (.cpp and .hpp) are fully new additions to the codebase.