Refactor code formatting and order of includes
This commit entails purely cosmetic changes. Code formatting was improved by adding spaces after commas and around operators for better readability and consistency. Additionally, the order of includes was rearranged in 'DrawCheckerboardQuad.hpp' for adherence to guidelines, along with minor tweaks such as adjustments of comments and removal of an excessive empty line. No changes were made to the functional aspects of the code.