Add EnableBlendingCommand to graphics engine
Enabled Blending Command was added to the graphics engine to enhance render flexibility. This included the creation of `EnableBlendingCommand.cpp` and `GLEnableBlending.hpp` files, and an update to the CMakeLists.txt file to reflect the resulting changes. This feature ensures smooth transitions between transparent and opaque objects and thus it could promote realistic visual appearances.