Add support for variable line width in drawLines method
The drawLines method in GLDrawCalls.hpp has been updated to support variable line widths. A new optional parameter, lineWidth, was added to this function's signature, allowing developers to specify the desired line width. If this parameter isn't provided, a default value of 1.0f will be used. This change provides more flexibility and customization options when drawing lines in graphics using the OpenGL framework. In order to apply this line width, the GL_LINE_WIDTH attribute is enabled before and disabled after the draw call.