Refactor grid rendering in hexen engine
Updated the rendering process of grids in the hexen engine. Introduced a "resize" method in DrawGridCommand to reduce redundancy and ensure that changes in grid size or unit size are correctly reflected in the rendering. Made explicit the linewidth for the grid and improved the grid vertices calculation. This increases the robustness, adaptability, and efficiency of the grid rendering, especially when manipulating the grid size or unit size. Also, tidied up unnecessary lines and better documenting the GridComponent functionalities.