Update drawLines function to accept line width parameter
Added 'lineWidth' parameter to drawLines function in DrawCalls.hpp. This allows users to customize the width of the lines drawn, providing more flexibility in rendering graphics. 'lineWidth' parameter is now passed to OpenGL drawLines call thus providing the ability to modify the line width at runtime in API supporting it.